requirements.txt
streamlit==1.50.0
tensorflow==2.20.0
torch==2.8.0
ultralytics==8.3.234
numpy==2.0.2
pandas==2.3.3
pillow==11.3.0
matplotlib==3.9.4
